backend/requirements.txt
asgiref==3.8.1 ; python_version >= "3.10" and python_version < "4.0"
certifi==2025.1.31 ; python_version >= "3.10" and python_version < "4.0"
cffi==1.17.1 ; python_version >= "3.10" and python_version < "4.0" and platform_python_implementation != "PyPy"
charset-normalizer==3.4.1 ; python_version >= "3.10" and python_version < "4.0"
cloudinary==1.42.2 ; python_version >= "3.10" and python_version < "4.0"
cryptography==44.0.0 ; python_version >= "3.10" and python_version < "4.0"
defusedxml==0.7.1 ; python_version >= "3.10" and python_version < "4.0"
django-cors-headers==4.7.0 ; python_version >= "3.10" and python_version < "4.0"
django-filter==24.3 ; python_version >= "3.10" and python_version < "4.0"
django==5.1.6 ; python_version >= "3.10" and python_version < "4.0"
djangorestframework-simplejwt==5.4.0 ; python_version >= "3.10" and python_version < "4.0"
djangorestframework==3.15.2 ; python_version >= "3.10" and python_version < "4.0"
djoser==2.3.1 ; python_version >= "3.10" and python_version < "4.0"
drf-nested-routers==0.94.1 ; python_version >= "3.10" and python_version < "4.0"
drf-yasg==1.21.8 ; python_version >= "3.10" and python_version < "4.0"
idna==3.10 ; python_version >= "3.10" and python_version < "4.0"
inflection==0.5.1 ; python_version >= "3.10" and python_version < "4.0"
oauthlib==3.2.2 ; python_version >= "3.10" and python_version < "4.0"
packaging==24.2 ; python_version >= "3.10" and python_version < "4.0"
pillow==11.1.0 ; python_version >= "3.10" and python_version < "4.0"
psycopg2-binary==2.9.10 ; python_version >= "3.10" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.10" and python_version < "4.0" and platform_python_implementation != "PyPy"
pyjwt==2.10.1 ; python_version >= "3.10" and python_version < "4.0"
python-decouple==3.8 ; python_version >= "3.10" and python_version < "4.0"
python3-openid==3.2.0 ; python_version >= "3.10" and python_version < "4.0"
pytz==2025.1 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.10" and python_version < "4.0"
requests-oauthlib==2.0.0 ; python_version >= "3.10" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.10" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.10" and python_version < "4.0"
social-auth-app-django==5.4.2 ; python_version >= "3.10" and python_version < "4.0"
social-auth-core==4.5.4 ; python_version >= "3.10" and python_version < "4.0"
sqlparse==0.5.3 ; python_version >= "3.10" and python_version < "4.0"
sslcommerz-lib==1.0 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version == "3.10"
tzdata==2025.1 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32"
uritemplate==4.1.1 ; python_version >= "3.10" and python_version < "4.0"
urllib3==2.3.0 ; python_version >= "3.10" and python_version < "4.0"
whitenoise==6.9.0 ; python_version >= "3.10" and python_version < "4.0"
